Move hint helper function to geopm_hint.h
