+++ service/src/Helper.cpp
@@ -20,7 +20,6 @@
 #include <algorithm>
 #include <sstream>
 #include <map>
-#include "geopm_hint.h"
 #include "geopm_field.h"
 #include "geopm/Exception.hpp"
 #include "config.h"
@@ -207,21 +206,6 @@ namespace geopm
         return result;
     }
 
-    void check_hint(uint64_t hint)
-    {
-        if ((hint & ~GEOPM_MASK_REGION_HINT) != 0ULL) {
-            throw Exception("Helper::" + std::string(__func__) +
-                            "(): invalid hint: " +
-                            geopm::string_format_hex(hint),
-                            GEOPM_ERROR_INVALID, __FILE__, __LINE__);
-        }
-        if (hint >= GEOPM_SENTINEL_REGION_HINT) {
-            throw Exception("Helper::" + std::string(__func__) +
-                            "(): hint out of range: " +
-                            geopm::string_format_hex(hint),
-                            GEOPM_ERROR_INVALID, __FILE__, __LINE__);
-        }
-    }
 
     std::function<std::string(double)> string_format_type_to_function(int format_type)
     {
